third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/contentSecurityPolicy/nonces/style-enforce-blocked-expected.txt

CONSOLE ERROR: line 20: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'nonce-abc'". Either the 'unsafe-inline' keyword, a hash ('sha256-wfwlzcPBVjaRRXbeco7e+YXOZ0p3kzYLthUXIyu4XOw='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 25: Refused to load the stylesheet 'http://127.0.0.1:8000/security/contentSecurityPolicy/style-set-red.css' because it violates the following Content Security Policy directive: "style-src 'nonce-abc'".

This is a testharness.js-based test.
PASS Incorrectly nonced style blocks generate reports. 
PASS Incorrectly nonced stylesheets do not load. 
Harness: the test ran to completion.
